mika#1444: pilot impl (dispatch-lib PR-create recovery, mika#1396) (#1490)
* docs(plans): groom mika issue#1444 (content-only, architect pending)


* refactor(mika#1259): extract evidence/ module from db.rs and agent.rs

Pure code relocation — no behavior changes. Moves evidence-domain code
into `crates/mika-agent/src/evidence/` per Foundation §6 partition:

- evidence/audit.rs: AuditEvent struct + 10 Database methods + 3 tests
  (from db.rs)
- evidence/guards.rs: fabrication-guard predicates, grounding-rule
  enforcement helpers, regex statics + 53 tests (from agent.rs)
- evidence/mod.rs: module doc-comment + pub re-exports

db.rs re-exports AuditEvent for backward compatibility with paginated
list methods (owned by #1445 dashboard_queries). rewind.rs import path
updated. All call sites in agent.rs use the new evidence::guards path.


---------
